[Chore] Address review feedback — rename norm/base.py to norm_base.py
Extend <module>_base.py convention to tileops/ops/norm/base.py.
