+++ tileops/ops/norm/__init__.py
@@ -1,12 +1,12 @@
 from .ada_layer_norm import AdaLayerNormFwdOp
 from .ada_layer_norm_zero import AdaLayerNormZeroFwdOp
-from .base import RowNormOp
 from .batch_norm import BatchNormBwdOp, BatchNormFwdOp
 from .fused_add_layer_norm import FusedAddLayerNormFwdOp
 from .fused_add_rms_norm import FusedAddRMSNormFwdOp
 from .group_norm import GroupNormFwdOp
 from .instance_norm import InstanceNormFwdOp
 from .layer_norm import LayerNormFwdOp
+from .norm_base import RowNormOp
 from .rms_norm import RMSNormFwdOp
 
 __all__: list[str] = [
